200khz update
refactored to go back to a single call to stepgen.
added a macro to make the step generation DRY.
reduced the step accumulator to 65536 (not really sure if this is a good
idea)
increased the core timer rate to 200khz (52% CPU for step generation on
4 Axis)
Made the appropriate changes to the HAL.
programmable stepwidth is still not supported.